tests: Fix failing test script
The test failure was a protocol implementation issue. See the simplified
command line:

    echo "abc" | nc server port

Here, nc proceeds as follows:

    1. Connects to server:port
    2. Reads "abc\n" from stdin
    3. Sends "abc\n" to the peer
    4. Detects EOF
    5. Closes the connection

So, nc won't receive the echo response from the server. Of course, nc
may concurrently receive data and write the data to stdout, but that
must be done within the very short time interval between the end of 3
and 5. So, the echo protocol would succeed in a very low probability.

In this commit, we replace nc with a custom client that works like this:

    echo "abc" | echoclient server:port

    1. Connects to server:port
    2. Reads "abc\n" (and detects EOF) from stdin
    3. Sends "abc\n" to the peer
    4. Receives data from the server (data should be "abc\n")
    5. Writes the data to stdout
    6. Closes the connection

The client knows the protocol, so it receives data before closing
connection. Now, the test succeeds.