server: Fix ping-pong to kill zombies
I thought the `deadline` argument was used by Gorilla to automatically
detect pong timeout, but it was not. Gorilla does receive PongMessage
from peer, but timeout must be handled by the programmer by setting
PongHandler.

https://pkg.go.dev/github.com/gorilla/websocket#hdr-Control_Messages

So, the previous code just sended ping to the channel and did not
checked for response or no-response from agent. Thus, the server could
not detect a zombie session until the TCP connection dies on the server
side.